Rename tests; Improve codestyle
